Moved array implementation into array.c/.h
